Use some more std::transform and std::remove_copy_if to modify / filter sequences.
Cleanup some more loops and whatnot.

Move has_foreground_actions from schedule_ready to arise

git-svn-id: https://simulationcraft.googlecode.com/svn/branches/mop@11459 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ engine/report/sc_report.hpp
@@ -12,9 +12,9 @@ struct chart {
   static std::string resource_color( int type );
 
   static std::string raid_downtime ( const std::vector<player_t*> &players_by_name, int print_styles );
-  static int raid_aps ( std::vector<std::string>& images, const sim_t*, const std::vector<player_t*>&, bool dps );
-  static int raid_dpet( std::vector<std::string>& images, const sim_t* );
-  static int raid_gear( std::vector<std::string>& images, const sim_t* );
+  static size_t raid_aps ( std::vector<std::string>& images, const sim_t*, const std::vector<player_t*>&, bool dps );
+  static size_t raid_dpet( std::vector<std::string>& images, const sim_t* );
+  static size_t raid_gear( std::vector<std::string>& images, const sim_t* );
 
   static std::string action_dpet        ( const player_t* );
   static std::string aps_portion        ( const player_t* );
